drm: Use atomic mode for setting planes
drmModeSetPlane() has long been marked legacy(outdated) and should not
be used anymore. So we switch to atomic mode, like it is already done
for the other drm commits.

Instead of the drmModeSetPlane we simply create an atomic mode request,
set the planes properties and commit that request.

With this commit we consequently use the drm atomic mode in the plugin now.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>