drm: Move OSD modesetting to Frame2Display
Depending on OsdShown which was set by VideoOsdClear() and VideoOsdDrawARGB()
we set the plane properties in Frame2Display() now. This saves us the
unnecessary atomic commit in the Clear() and DrawARGB() functions. This is
possible, because we do an atomic commit in Frame2Display() anyway and just
add the plane properties now.
Clear() and DrawARGB() just do the memcpy/memset and set some variables,
which we need later in Frame2Display().