drm: Introduce dirty flag
The move of the property setting to Frame2Display() causes the properties
to be set every frame in Frame2Display().
We only need to do so, if we had any recent activity in VideoOsdClear() or
VideoOsdDrawARGB(). Introduce a dirty flag which belongs to the osd buffer,
set it in Clear/Draw and clear it after the new properties are set.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>