GL: Fix y-axis issue
We have to take of gl's origin siiting in the bottom left corner.
The GL code flips this withing rendering to RenderBuffer and does a
framebuffer blit (render to a texture) afterwards.
Within GLES we have to deal with y-inverted textures in the RenderBuffer
command and do the the flip within rendering to the output buffer,
so we insert the necessary y flips.

This fixes issues, which can be seen with skins that render multiple
pixmaps into one OSD e.g. SkinElchiHD. This commit should fix it.