drm: Rework SetPropertyRequest
It's probably expensive to iterate to the properties just before the atomic commit.
So once we have the plane_id, we fill an array with its properties and can
just read this array later.

We introduce a struct with the planes' properties and move the plane_id into this struct.