GL: Fix colors
While OpenGL's glTexImage2D had the possibility for a channels swap with
the format parameters, this can't be done in OpenGL/ES.
During my code conversion from GL to GL/ES i somehow ignored that.
We solve this with a copy of the texture fragment shader, which
swaps the R and B value. We use this within cOglCmdDrawTexture and
cOglCmdDrawImage which are only triggered by Pixmap->DrawImage() and
Pixmap->DrawBitmap() which should be right.

The old shader, which doesn't swap the channels is still used for our
functions which render the pixmaps (as textures) to the framebuffer
and to the output framebuffer in the end.