drm: Fix FilterHandlerThread endless loop
If av_buffersink_get_frame() returns "invalid argument" (not EAGAIN or EOF) we
run into an endless loop. Though we might be able to feed av_buffersrc_add_frame_flags
with a NULL frame to trigger EOF, av_buffersink_get_frame can't get that EOF because
we always get the same error.
We need to reset the filtergraph but can't exit the FilterHandlerThread the usual way.
So we simply add another label, free the frames and end the thread. It's restarted the
next time we want to render a frame.